fpc: .pyenv missing off default prompt
I adjusted the original change to include `/.pyenv`. I think it would be better to have a literal `$HOME` displayed if it can be, that's the point of the two lines vs one.  For example, if someone wants to use `~/Envs`, the output with the substitution version would be:

```bash
export PATH="$HOME/Envs/bin:$PATH"
```

with the one line, it would be:

```bash
export PATH="/home/yyuu/Envs/bin:$PATH"
```

which is slightly less aesthetically pleasing.